requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
docx2pdf==0.1.8
et_xmlfile==2.0.0
Flask==3.1.2
fpdf==1.7.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
PyMuPDF==1.26.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
pytz==2025.2
pywin32==311
six==1.17.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
xlsxwriter==3.2.5
